+++ meta/src/tcpsocket.h
@@ -35,7 +35,7 @@ ssize_t tcpsocket_read(tcpsocket p, char *buf, size_t count,
     __attribute__((nonnull))
     __attribute__((warn_unused_result));
 
-status_t tcpsocket_poll_for(tcpsocket p, int timeout, int polltype)
+status_t tcpsocket_poll_for(tcpsocket p, int timeout, short polltype)
     __attribute__((nonnull))
     __attribute__((warn_unused_result));
